refactor: remove unused layout component and enhance support page API requests
